[BUGFIX] Move simulate action from pagetitlesim to pagetitle
 This correction never affected a production version of MetaSeo
 (regression  was in develop branch only, introduced
 with a78f7304)

Issue #160
